create coordinators when a cluster node is up
